requirements.txt
numpy
scipy
cupy-cuda11x
networkx
matplotlib